Only check for timeout, not failed requests
If we check for both, then it seems that both get triggered whenever there is a
timeout. The code for the server only ever respnods with 200 OK, so it doesn't
make sense that a request would fail anyway. If it times out, then we do the
right thing and potentially inform the user.