Automatically generate HTML when server starts
Michael pulled the latest changes from the repo and tried running the server,
but it wasn't behaving correctly. The problem was that he hadn't generated the
templates (I had changed the template text files). To prevent this from
hapenning in the future, I made main.py automatically run the template script
when it starts up. I moved it to src/ because of this, and renamed it.