Add controls/drawing toggle & more
I added a toggle button to go between the controls page and the drawing page.
Both live in the same HTML file, but the JS simply sets one of them to have
'display: none'. I hooked Justin's JS to work with the web app, and I started
implementing the point-sending process. There is still some more work to be done
there. I also did a lot of renaming. We tested the program once, and the angle
calculations appear to be way off.