Don't try turning small angles; make first dot red
The robot always turns before driving along the next segment, even if the angle
it has to turn is very small. If it tries turning a 0.1 degree angle, for
example, it will start turning and then stop as fast as possible, which results
in a much too large rotation due to Bluetooth latency and other factors. I added
a check to make the robot skip the rotation phase if the rotation is less than a
certain number of degrees (controlled by a parameter).

It is easy to forget which dot you placed first, and knowing this is crucial for
placing the robot on the paper (if you put it in the wrong position, it might
drive off the paper), so I made `drawing.js` draw the first dot with a red fill.