Refactor util functions into separate module
I removed other.py -- this was originally supposed to be a quick prototype of
the "other" project which became Tracie, and I should have deleted it a long
time ago. I refactored a bunch of mathematical functions into a new module in
util.py. This includes degrees/radians conversions, 2d distance, related
equivalent angles, and list averages. I updated all the program modules to use
the util module. Finally, I added an angle->time operation as the inverse to
time->angle since there was only an inverse for distance before, and it felt
asymmetrical.