Hide draw button in Avoider; don't send < 2 points
Now, when you switch to the Avoider program, the draw/controls toggle button
disappears, since it makes no sense to draw on the canvas and send points in
that mode. Also, now the web app doesn't even try to send points if there aren't
enough, preventing some errors in the server where it tries to access the first
element of the empty list.