Add parameter help page
We used to only have two "views". I added a third one, the 'param-help' view. I
generalized the code to deal with multiple pages when I did this (no more
'toggle' functions). There is now a function to switch to a given view.

The help view displays a definition list of the short codes and their full
names. (I guess I could also include full sentence descriptions in the Python
code and display these as well, but I'm too lazy to do that.) I spent a while
styling it (using 'display: table' once again). I created a new POST request for
the current program's shortcode dictionary. When you click on the "Help" button,
the web app requests this, turns it into HTML with the shortcodes sorted
alphabetically, and replaces the contents of the help page. This all happens
instantly as the controls view is hidden and the 'param-help' view is revealed.