Fix button enable states; start working on Trace
The buttons in the drawing view were disabled at the appropriate times, but you
could still click them and their actions would be carried out. Rather than
adding an isEnabled check to each of these functions, I refactored everything
into a single btnAction function which checks if the button is enabled and if
so, looks up the button's action in a dictionary and calls it.

I realized that the "Clear" button disables itself in the drawing view but not
in the controls view, so I made the controls "Clear" button also disable itself
for consistency. Finally, I started working on the Trace button in the drawing
view. It now activate and deactives like the "DEL" button, but does nothing
other than that. I will have to make a new type of 'long:' request, I think.