Fix bug where latest point_scale is not used
If you changed the point_scale, it had no effect until you sent a new set of
points. This is how I wrote it, but I wasn't intending this behaviour. If you
change point_scale before starting the program (but after sending the points),
or even in the middle of the execution of the program, it should use it. Now,
instead of scaling the points in the transform method, the distance is simply
multiplied by point_scale in the set_drive_time function.