More work on server/client timeout
The gevent queue timeout is good, but I forgot that it raises an exception when
it times out. When I catch the exception, I have to return something, so I
decided to return None and then took this into consideration in the server's
handle_post method. Now, if a command returns None (instead of a status string),
the server will send a zero-length message with status 204 NO CONTENT. I'm not
sure if this is the best header to use, but it doesn't really matter.

In post function of controls.js, I added another callback for failing (separate
from client timeout). This is called if there is a response, but its header is
not 200 OK. The updateStatus function simply calls itself again if this happens,
just like what it does when it times out. The other posting functions add a
message to the console indicating the failure and its status code.

I decided to leave the client timeout at 30 seconds and change the server
timeout to 31 seconds. Now, the client will give up just before the server does.
The server will respond with 204, but the client won't get that because it is no
longer listening for it. I could do it the other way around and have the server
give up one second before the client; in this case, the client would never time
out (in these specific circumstances) because the server would always response
with 204 in time (or 200 some time earlier). I don't think either option has an
advantage over the other, so I'm leaving it at client-30 and server-31 for now.