Fix more issues with long-polling & timeouts
I keep thinking I am done with this, but then a new issue appears! When I
started the Tracie program, sometimes it would not should the first one or two
statuses for some reason. I figured something was wrong with the queue access --
somehow those statuses were getting lost. I realized that there IS an advantage
to making the server time out first, instead of the other way around (which is
what I did in the last commit), because if the client gives up, the server can
still try to return something from the queue, and then that item is lost
forever.

So I changed it to server-25 and client-30, but that didn't solve my problems. I
changed the client timeout to be very small for testing purposes, and at some
point my browser started crashing. The I noticed that the errors in the Web
console were growing exponentially! This is because each failed or timed out
status update starts a new one -- but for some reason, a timeout causes
ontimeout to be called AND the request to be completed with status 0 (not 204),
so each timeout triggered two new status updates.

I added a check for a zero status. I have no idea why those happen, but now they
will be ignored. The server times out before the client and responds with 204 NO
CONTENT, and the client sees this and initiates a new request without timing
out. One good thing about this is that there aren't any errors in the JS console
-- I probably shouldn't have been using timeouts for things that were supposed
to happen before. I think everything is finally good now.

The bug I mentioned earlier, about statuses being missing, was very hard to
reproduce, even before these changes. I think its gone, but I can't be sure.

I also made the Controller's start method use start_later on its Greenlet
instead of an immediate start, because the first status of the program was
beating the "program resumed" message to the console. That seems to have fixed
it. It would probably be better if I had ALL messages in a queue, guaranteed to
appear in the console in the order that they occurred, but that would be a lot
of work for little gain at this point.