Use 'points:' prefix; resend points on switch
Before I was assuming that any client command beginning with '[' was JSON data
for the points. Now, a prefix of 'points:' is used for consistency with
everythign else. Also, when you switch to the Tracie program, the points are
resent to the server (before you had to go to the drawing page and back, which
was not intuitive).